Support TIME in off-heap column vectors
